requirements.txt
flask>=3.0
flask-login>=0.6
flask-sqlalchemy>=3.1
python-dotenv>=1.0
google-genai>=1.0
pdfplumber>=0.11
python-docx>=1.1
werkzeug>=3.0
gunicorn>=21.0
playwright>=1.40
apscheduler>=3.10
beautifulsoup4>=4.12
